requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
faiss-cpu
transformers
scikit-learn
numpy
pandas
tiktoken
python-dotenv
regex